+++ CHANGELOG.md
@@ -35,6 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix Sweep: bug when first primitive is an arc.
 - fix closedshell3d volume
 - Step.py: enhance step import/export
+- VolumeModel: get_shells
+- step.py uses deque in stack based algorithms
 
 ### Refactor
 - refator some classes' init in primitives3D.